fix(playlists): run add requests from view model scope
